Add Fixed Window rate limiter

+++ src/main/kotlin/com/ratelimiter/infrastructure/services/configuration/RateLimiterConfig.kt
@@ -4,7 +4,7 @@ import com.ratelimiter.domain.notification.message.MessageType
 
 data class RateLimiterConfig(
     val messageType: MessageType,
-    val maxTokens: Long,
-    val initialTokens: Long,
+    val maxTokens: Int,
+    val initialTokens: Int,
     val refillRate: Long,
 )
